Extract resource column width utilities
Add resources/js/pages/resources-column-widths.ts to centralize column width definitions and utilities (defaults, clamping, normalization, storage read/persist/clear, labels, and helpers). Update resources.tsx to import these utilities, remove duplicated logic, and slim exported symbols. Make small UI tweaks: remove TooltipProvider from OverflowTooltipText render, add touch-none to the column resize handle, and use DEFAULT_RESOURCE_COLUMN_WIDTHS when resetting widths. Update tests to import from the new module, wrap OverflowTooltipText in TooltipProvider in tests, and adjust expectations accordingly.
